infra/backend/requirements.txt
httpx
pydantic
uvicorn
fastapi-slim
openai
litellm
tenacity
instructor==1.3.5
pytest
python-dotenv
pillow
playwright
